+++ packages/core/src/index.d.ts
@@ -228,6 +228,7 @@ export enum RotateDirection {
 }
 
 export enum ScrollMode {
+    Page = 'Page',
     Horizontal = 'Horizontal',
     Vertical = 'Vertical',
     Wrapped = 'Wrapped',