+++ packages/core/src/structs/ScrollMode.ts
@@ -7,6 +7,7 @@
  */
 
 export enum ScrollMode {
+    Page = 'Page',
     Horizontal = 'Horizontal',
     Vertical = 'Vertical',
     Wrapped = 'Wrapped',