+++ packages/scroll-mode/src/index.ts
@@ -7,6 +7,7 @@
  */
 
 export { HorizontalScrollingIcon } from './HorizontalScrollingIcon';
+export { PageScrollingIcon } from './PageScrollingIcon';
 export * from './scrollModePlugin';
 export type { SwitchScrollModeProps } from './SwitchScrollMode';
 export type { RenderSwitchScrollModeProps } from './types/RenderSwitchScrollModeProps';